try float16

+++ run_qlora.py
@@ -32,7 +32,7 @@ def main():
             quantization_config=BitsAndBytesConfig(
                 load_in_4bit=True,
                 llm_int8_has_fp16_weight=False,
-                bnb_4bit_compute_dtype=torch.float32,
+                bnb_4bit_compute_dtype=torch.float16,
                 bnb_4bit_use_double_quant=False,
                 bnb_4bit_quant_type='nf4',
             ),
